programming_paradigm

+++ programming_paradigm/robust_division_calculator.py
@@ -1,6 +1,6 @@
 def safe_divide(numerator, denominator):
     try:
-        result = int(numerator) / int(denominator)
+        result = float(numerator) / float(denominator)
         print(f"The result of {numerator} / {denominator} is {result}")
     except ZeroDivisionError:
         print(f"Hello your numerator {numerator} is Zero and is invalid")
